flatpak: Simplify TrackerSparqlConnection creation
The Tracker D-Bus services are not part of the Flatpak bundle and are
accessed from the host operating system. This means that the Tracker
database cannot be accessed "directly", which causes:
  Tracker-WARNING **: Falling back to bus backend, the direct backend
  failed to initialize: Could not open sqlite3 database:
  '/home/rishi/.var/app/org.gnome.Photos/cache/tracker/meta.db':
  unable to open database file

The database needs to be accessed via D-Bus. So far, this was being
handled by the fallback error-handling code paths.