flatpak: Switch to Meson
The nightly Flatpak has been using Meson for a while. Hence, this is a
good time to switch Flathub.

The Autotools build didn't use --disable-dogtail, even though the SDK
doesn't have Dogtail. It only worked because configure.ac doesn't
actually check for the presence of Dogtail. The Meson build doesn't
check for it either, but merely to imitate the Autotools build.

It's better to spell out exactly what's going on.

https://github.com/flathub/org.gnome.Photos/pull/9