flatpak: Don't use a non-existent build flag
There has never ever been a --disable-documentation build flag, yet
somehow it has been present in the Flatpak manifest since it was first
added to the gnome-apps-nightly repository.

GNOME Online Accounts has a --disable-documentation flag to disable
generating the man page for goa-daemon(8), which is already covered by
the --disable-backend flag. So, it might have been a typo while writing
the manifest.