Update grilo, grilo-plugins, add totem-pl-parser
grilo 0.3.10 -> 0.3.12: Fix multiple build issues and update translations.
Also, set -Denable-grl-pls to true, as the new version of grilo-plugins
needs it to compile. Unfortunately, this means we have to add
totem-pl-parser to the manifest, as now grilo requires it.

grilo-plugins 0.3.10 -> 0.3.11: Fix crash if chromaprint plugin isn't
installed, this applies to us (#67), and update translations.